Change ShroudCopyStringAndFree to size_t c_var_len
It was long.  SHROUD_array%len is passed directly
to this function (and it is size_t).

+++ tests/example/wrapExClass1.cpp
@@ -40,6 +40,7 @@
 //
 // #######################################################################
 #include "wrapExClass1.h"
+#include <cstddef>
 #include <cstring>
 #include <stdlib.h>
 #include <string>
@@ -64,7 +65,7 @@ static void ShroudStrCopy(char *a, int la, const char *s)
 
 // Copy the std::string in context into c_var.
 // Called by Fortran to deal with allocatable character.
-void AA_ShroudCopyStringAndFree(USE_SHROUD_array *data, char *c_var, long c_var_len) {
+void AA_ShroudCopyStringAndFree(USE_SHROUD_array *data, char *c_var, size_t c_var_len) {
     const char *cxx_var = data->addr.ccharp;
     size_t n = c_var_len;
     if (data->len < n) n = data->len;
